# server/swagger/latest.yaml
openapi: 3.0.1
info:
  title: IPNS Xumm Api
  contact:
    email: dangell@transia.co
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
externalDocs:
  description: What is IPNS Xumm API
  url: https://ipns.transia.co
servers:
- url: https://ipns-api-3uzuax5sua-uc.a.run.app/v1
tags:
- name: ipns
  description: IPNS API Endpoints
paths:
  /ipns/pin:
    post:
      tags:
      - ipns
      summary: Pin an IPNS Json
      operationId: ipns_pin_json
      requestBody:
        description: Json to publish
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IPNSPinRequest'
        required: true
      responses:
        "200":
          description: IPNS Pin Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPNSPinResponse'
        "400":
          description: Invalid Params supplied
        "405":
          description: Validation exception
      security:
      - bearerAuth:
        - token_info
      x-openapi-router-controller: server.controllers.ipns_controller
  /ipns/publish:
    post:
      tags:
      - ipns
      summary: Publish an IPNS CID
      operationId: ipns_publish
      requestBody:
        description: CID to publish
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IPNSPublishRequest'
        required: true
      responses:
        "200":
          description: IPNS Publish Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPNSPublishResponse'
        "400":
          description: Invalid Params supplied
        "405":
          description: Validation exception
      security:
      - bearerAuth:
        - token_info
      x-openapi-router-controller: server.controllers.ipns_controller
  /ipns/resolve:
    post:
      tags:
      - ipns
      summary: Resolve an IPNS Name
      operationId: ipns_resolve
      requestBody:
        description: Name to resolve
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IPNSResolveRequest'
        required: true
      responses:
        "200":
          description: IPNS Resolve Successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IPNSResolveResponse'
        "400":
          description: Invalid Params supplied
        "405":
          description: Validation exception
      security:
      - bearerAuth:
        - token_info
      x-openapi-router-controller: server.controllers.ipns_controller
components:
  schemas:
    MessageResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      example:
        code: 0
        type: type
        message: message
    IPNSPinRequest:
      required:
      - json
      type: object
      properties:
        json:
          type: object
          additionalProperties: {}
    IPNSPinResponse:
      required:
      - name
      type: object
      properties:
        name:
          type: string
    IPNSPublishRequest:
      required:
      - cid
      type: object
      properties:
        cid:
          type: string
    IPNSPublishResponse:
      required:
      - name
      type: object
      properties:
        name:
          type: string
    IPNSResolveRequest:
      required:
      - name
      type: object
      properties:
        name:
          type: string
    IPNSResolveResponse:
      required:
      - cid
      type: object
      properties:
        cid:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: server.controllers.authentication_controller.bearer_auth